+++ RELEASE_NOTES.md
@@ -17,6 +17,7 @@
 * GMB-1164: Add an initial pipeline ID info to workResult
 * Property update : Added MAPR ticket file location in g2mailboxservice properties at all environment
 * GQ-305: Add support for Kafka consumer thread pool bleed
+* GMB-1192 - Add QS consumers for inbound file and processor retry queues
 
 5.4.9
 ---------------